Switching dataset double to float

+++ src/Analyzer.cpp
@@ -85,26 +85,26 @@ int Analyzer::falsePositive(Class c) {
     return confusionMatrix[nClasses][c] - confusionMatrix[c][c];
 }
 
-double Analyzer::overall(Class c) {
-    return (double) (truePositive(c) + trueNegative(c)) / (double) confusionMatrix[nClasses][nClasses];
+float Analyzer::overall(Class c) {
+    return (float) (truePositive(c) + trueNegative(c)) / (float) confusionMatrix[nClasses][nClasses];
 }
 
-double Analyzer::precision(Class c) {
-    return (double) truePositive(c) / (double) (truePositive(c) + falsePositive(c));
+float Analyzer::precision(Class c) {
+    return (float) truePositive(c) / (float) (truePositive(c) + falsePositive(c));
 }
 
-double Analyzer::negativePredictedValue(Class c) {
-    return (double) trueNegative(c) / (double) (trueNegative(c) + falseNegative(c));
+float Analyzer::negativePredictedValue(Class c) {
+    return (float) trueNegative(c) / (float) (trueNegative(c) + falseNegative(c));
 }
 
-double Analyzer::sensitivity(Class c) {
-    return (double) truePositive(c) / (double) (truePositive(c) + falseNegative(c));
+float Analyzer::sensitivity(Class c) {
+    return (float) truePositive(c) / (float) (truePositive(c) + falseNegative(c));
 }
 
-double Analyzer::specificity(Class c) {
-    return (double) trueNegative(c) / (double) (trueNegative(c) + falsePositive(c));
+float Analyzer::specificity(Class c) {
+    return (float) trueNegative(c) / (float) (trueNegative(c) + falsePositive(c));
 }
 
-double Analyzer::fScore(Class c) {
+float Analyzer::fScore(Class c) {
     return 2.0 * precision(c) * sensitivity(c) / (precision(c) + sensitivity(c));
 }
